#4 Move AddBankAccount to own namespace
